move txn manager out of batcher
