+++ common/mock/txn_manager.go
@@ -3,17 +3,17 @@ package mock
 import (
 	"context"
 
-	"github.com/Layr-Labs/eigenda/disperser/batcher"
+	"github.com/Layr-Labs/eigenda/common"
 	"github.com/stretchr/testify/mock"
 )
 
 type MockTxnManager struct {
 	mock.Mock
 
-	Requests []*batcher.TxnRequest
+	Requests []*common.TxnRequest
 }
 
-var _ batcher.TxnManager = (*MockTxnManager)(nil)
+var _ common.TxnManager = (*MockTxnManager)(nil)
 
 func NewTxnManager() *MockTxnManager {
 	return &MockTxnManager{}
@@ -21,13 +21,13 @@ func NewTxnManager() *MockTxnManager {
 
 func (b *MockTxnManager) Start(ctx context.Context) {}
 
-func (b *MockTxnManager) ProcessTransaction(ctx context.Context, req *batcher.TxnRequest) error {
+func (b *MockTxnManager) ProcessTransaction(ctx context.Context, req *common.TxnRequest) error {
 	args := b.Called()
 	b.Requests = append(b.Requests, req)
 	return args.Error(0)
 }
 
-func (b *MockTxnManager) ReceiptChan() chan *batcher.ReceiptOrErr {
+func (b *MockTxnManager) ReceiptChan() chan *common.ReceiptOrErr {
 	args := b.Called()
-	return args.Get(0).(chan *batcher.ReceiptOrErr)
+	return args.Get(0).(chan *common.ReceiptOrErr)
 }
