+++ disperser/batcher/metrics.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/Layr-Labs/eigenda/common"
 	"github.com/Layr-Labs/eigenda/core"
 	"github.com/Layr-Labs/eigenda/disperser"
 	"github.com/Layr-Labs/eigensdk-go/logging"
@@ -35,14 +36,6 @@ type EncodingStreamerMetrics struct {
 	EncodedBlobs *prometheus.GaugeVec
 }
 
-type TxnManagerMetrics struct {
-	Latency  *prometheus.SummaryVec
-	GasUsed  prometheus.Gauge
-	SpeedUps prometheus.Gauge
-	TxQueue  prometheus.Gauge
-	NumTx    *prometheus.CounterVec
-}
-
 type FinalizerMetrics struct {
 	NumBlobs               *prometheus.CounterVec
 	LastSeenFinalizedBlock prometheus.Gauge
@@ -55,7 +48,7 @@ type DispatcherMetrics struct {
 
 type Metrics struct {
 	*EncodingStreamerMetrics
-	*TxnManagerMetrics
+	*common.TxnManagerMetrics
 	*FinalizerMetrics
 	*DispatcherMetrics
 
@@ -88,46 +81,7 @@ func NewMetrics(httpPort string, logger logging.Logger) *Metrics {
 		),
 	}
 
-	txnManagerMetrics := TxnManagerMetrics{
-		Latency: promauto.With(reg).NewSummaryVec(
-			prometheus.SummaryOpts{
-				Namespace:  namespace,
-				Name:       "txn_manager_latency_ms",
-				Help:       "transaction confirmation latency summary in milliseconds",
-				Objectives: map[float64]float64{0.5: 0.05, 0.9: 0.01, 0.95: 0.01, 0.99: 0.001},
-			},
-			[]string{"stage"},
-		),
-		GasUsed: promauto.With(reg).NewGauge(
-			prometheus.GaugeOpts{
-				Namespace: namespace,
-				Name:      "gas_used",
-				Help:      "gas used for onchain batch confirmation",
-			},
-		),
-		SpeedUps: promauto.With(reg).NewGauge(
-			prometheus.GaugeOpts{
-				Namespace: namespace,
-				Name:      "speed_ups",
-				Help:      "number of times the gas price was increased",
-			},
-		),
-		TxQueue: promauto.With(reg).NewGauge(
-			prometheus.GaugeOpts{
-				Namespace: namespace,
-				Name:      "tx_queue",
-				Help:      "number of transactions in transaction queue",
-			},
-		),
-		NumTx: promauto.With(reg).NewCounterVec(
-			prometheus.CounterOpts{
-				Namespace: namespace,
-				Name:      "tx_total",
-				Help:      "number of transactions processed",
-			},
-			[]string{"state"},
-		),
-	}
+	txnManagerMetrics := common.NewTxnManagerMetrics(namespace, reg)
 
 	finalizerMetrics := FinalizerMetrics{
 		NumBlobs: promauto.With(reg).NewCounterVec(
@@ -170,7 +124,7 @@ func NewMetrics(httpPort string, logger logging.Logger) *Metrics {
 
 	metrics := &Metrics{
 		EncodingStreamerMetrics: &encodingStreamerMetrics,
-		TxnManagerMetrics:       &txnManagerMetrics,
+		TxnManagerMetrics:       txnManagerMetrics,
 		FinalizerMetrics:        &finalizerMetrics,
 		DispatcherMetrics:       &dispatcherMatrics,
 		Blob: promauto.With(reg).NewCounterVec(
@@ -312,26 +266,6 @@ func (g *Metrics) Start(ctx context.Context) {
 func (e *EncodingStreamerMetrics) UpdateEncodedBlobs(count int, size uint64) {
 	e.EncodedBlobs.WithLabelValues("size").Set(float64(size))
 	e.EncodedBlobs.WithLabelValues("number").Set(float64(count))
-}
-
-func (t *TxnManagerMetrics) ObserveLatency(stage string, latencyMs float64) {
-	t.Latency.WithLabelValues(stage).Observe(latencyMs)
-}
-
-func (t *TxnManagerMetrics) UpdateGasUsed(gasUsed uint64) {
-	t.GasUsed.Set(float64(gasUsed))
-}
-
-func (t *TxnManagerMetrics) UpdateSpeedUps(speedUps int) {
-	t.SpeedUps.Set(float64(speedUps))
-}
-
-func (t *TxnManagerMetrics) UpdateTxQueue(txQueue int) {
-	t.TxQueue.Set(float64(txQueue))
-}
-
-func (t *TxnManagerMetrics) IncrementTxnCount(state string) {
-	t.NumTx.WithLabelValues(state).Inc()
 }
 
 func (f *FinalizerMetrics) IncrementNumBlobs(state string) {
